[LUM] add Username

+++ src/main/java/ch/fhnw/lems/persistence/LoginHistoryRepository.java
@@ -11,5 +11,5 @@ import ch.fhnw.lems.dto.LoginHistory;
 @Repository
 public interface LoginHistoryRepository extends JpaRepository<LoginHistory, Long> {
 		@Query(value="select id from login_history where user_id =:user_id order by login DESC LIMIT 1", nativeQuery = true)
-		Integer lastUserLogin(@Param("user_id") Long user_id);
+		Long lastUserLogin(@Param("user_id") Long user_id);
 }
